requirements.txt
defusedxml==0.5.0
Django==1.9
django-browserid==2.0.2
django-classy-tags==0.8.0
django-crispy-forms==1.6.1
django-filter==1.0.1
django-markdown-deux==1.0.5
django-missing==0.1.18
django-pagedown==0.1.1
django-rest-framework-mongoengine==3.3.1
django-sekizai==0.10.0
djangorestframework==3.5.4
djangorestframework-xml==1.3.0
Markdown==2.6.8
markdown2==2.3.3
mechanize==0.2.5
mongoengine==0.11.0
oauthlib==2.0.1
pymongo==3.4.0
requests==2.13.0
requests-oauthlib==0.8.0
six==1.10.0
tweepy==3.5.0